- per offline feedback, move the code that determines the CMK error message from Core and into CosmosDb.  This removes any Cosmos-related knownledge from Core
- Cosmos DB has published documentation on BYOK substatus codes: https://docs.microsoft.com/en-us/rest/api/cosmos-db/http-status-codes-for-cosmosdb.  Updating Resources to better aligned with the Cosmos wording.
