+++ src/Microsoft.Health.Fhir.Core/Exceptions/CustomerManagedKeyException.cs
@@ -3,57 +3,18 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-using System;
 using Microsoft.Health.Fhir.Core.Models;
 
 namespace Microsoft.Health.Fhir.Core.Exceptions
 {
     public class CustomerManagedKeyException : FhirException
     {
-        public CustomerManagedKeyException(int? subStatusCode)
+        public CustomerManagedKeyException(string message)
         {
-            string errorMessage = Resources.CmkDefaultError;
-            if (subStatusCode.HasValue && Enum.IsDefined(typeof(KnownCosmosDbCmkSubStatusValue), subStatusCode))
-            {
-                switch ((KnownCosmosDbCmkSubStatusValue)subStatusCode.Value)
-                {
-                    case KnownCosmosDbCmkSubStatusValue.AadClientCredentialsGrantFailure:
-                        errorMessage = Resources.AadClientCredentialsGrantFailure;
-                        break;
-                    case KnownCosmosDbCmkSubStatusValue.AadServiceUnavailable:
-                        errorMessage = Resources.AadServiceUnavailable;
-                        break;
-                    case KnownCosmosDbCmkSubStatusValue.KeyVaultAuthenticationFailure:
-                        errorMessage = Resources.KeyVaultAuthenticationFailure;
-                        break;
-                    case KnownCosmosDbCmkSubStatusValue.KeyVaultKeyNotFound:
-                        errorMessage = Resources.KeyVaultKeyNotFound;
-                        break;
-                    case KnownCosmosDbCmkSubStatusValue.KeyVaultServiceUnavailable:
-                        errorMessage = Resources.KeyVaultServiceUnavailable;
-                        break;
-                    case KnownCosmosDbCmkSubStatusValue.KeyVaultWrapUnwrapFailure:
-                        errorMessage = Resources.KeyVaultWrapUnwrapFailure;
-                        break;
-                    case KnownCosmosDbCmkSubStatusValue.InvalidKeyVaultKeyUri:
-                        errorMessage = Resources.InvalidKeyVaultKeyUri;
-                        break;
-                    case KnownCosmosDbCmkSubStatusValue.InvalidInputBytes:
-                        errorMessage = Resources.InvalidInputBytes;
-                        break;
-                    case KnownCosmosDbCmkSubStatusValue.KeyVaultInternalServerError:
-                        errorMessage = Resources.KeyVaultInternalServerError;
-                        break;
-                    case KnownCosmosDbCmkSubStatusValue.KeyVaultDnsNotResolved:
-                        errorMessage = Resources.KeyVaultDnsNotResolved;
-                        break;
-                }
-            }
-
             Issues.Add(new OperationOutcomeIssue(
                     OperationOutcomeConstants.IssueSeverity.Error,
                     OperationOutcomeConstants.IssueType.Forbidden,
-                    errorMessage));
+                    message));
         }
     }
 }
