+++ src/Microsoft.Health.Fhir.CosmosDb/Features/Storage/CosmosResponseProcessor.cs
@@ -16,7 +16,6 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Primitives;
 using Microsoft.Health.Abstractions.Exceptions;
 using Microsoft.Health.CosmosDb.Features.Storage;
-using Microsoft.Health.Fhir.Core.Exceptions;
 using Microsoft.Health.Fhir.Core.Features.Context;
 using Microsoft.Health.Fhir.CosmosDb.Features.Metrics;
 
@@ -78,7 +77,7 @@ namespace Microsoft.Health.Fhir.CosmosDb.Features.Storage
                     int? subStatusValue = dce.GetSubStatusValue();
                     if (subStatusValue.HasValue && Enum.IsDefined(typeof(KnownCosmosDbCmkSubStatusValue), subStatusValue))
                     {
-                        throw new Core.Exceptions.CustomerManagedKeyException(subStatusValue);
+                        throw new Core.Exceptions.CustomerManagedKeyException(GetCustomerManagedKeyErrorMessage(subStatusValue.Value));
                     }
                 }
             }
@@ -153,6 +152,47 @@ namespace Microsoft.Health.Fhir.CosmosDb.Features.Storage
             {
                 _logger.LogCritical(ex, "Unable to publish CosmosDB metric.");
             }
+        }
+
+        private string GetCustomerManagedKeyErrorMessage(int subStatusCode)
+        {
+            string errorMessage = Resources.CmkDefaultError;
+
+            switch ((KnownCosmosDbCmkSubStatusValue)subStatusCode)
+            {
+                case KnownCosmosDbCmkSubStatusValue.AadClientCredentialsGrantFailure:
+                    errorMessage = Resources.AadClientCredentialsGrantFailure;
+                    break;
+                case KnownCosmosDbCmkSubStatusValue.AadServiceUnavailable:
+                    errorMessage = Resources.AadServiceUnavailable;
+                    break;
+                case KnownCosmosDbCmkSubStatusValue.KeyVaultAuthenticationFailure:
+                    errorMessage = Resources.KeyVaultAuthenticationFailure;
+                    break;
+                case KnownCosmosDbCmkSubStatusValue.KeyVaultKeyNotFound:
+                    errorMessage = Resources.KeyVaultKeyNotFound;
+                    break;
+                case KnownCosmosDbCmkSubStatusValue.KeyVaultServiceUnavailable:
+                    errorMessage = Resources.KeyVaultServiceUnavailable;
+                    break;
+                case KnownCosmosDbCmkSubStatusValue.KeyVaultWrapUnwrapFailure:
+                    errorMessage = Resources.KeyVaultWrapUnwrapFailure;
+                    break;
+                case KnownCosmosDbCmkSubStatusValue.InvalidKeyVaultKeyUri:
+                    errorMessage = Resources.InvalidKeyVaultKeyUri;
+                    break;
+                case KnownCosmosDbCmkSubStatusValue.InvalidInputBytes:
+                    errorMessage = Resources.InvalidInputBytes;
+                    break;
+                case KnownCosmosDbCmkSubStatusValue.KeyVaultInternalServerError:
+                    errorMessage = Resources.KeyVaultInternalServerError;
+                    break;
+                case KnownCosmosDbCmkSubStatusValue.KeyVaultDnsNotResolved:
+                    errorMessage = Resources.KeyVaultDnsNotResolved;
+                    break;
+            }
+
+            return errorMessage;
         }
     }
 }
